requirements.txt
alembic==1.14.0
annotated-types==0.7.0
anyio==4.6.2.post1
APScheduler==3.11.0
blinker==1.9.0
certifi==2024.8.30
charset-normalizer==3.4.0
click==8.1.7
DateTime==5.5
distro==1.9.0
dominate==2.9.1
Flask==3.1.0
Flask-Bootstrap==3.3.7.1
Flask-Migrate==4.0.7
Flask-Script==2.0.6
Flask-SQLAlchemy==3.1.1
Flask-WTF==1.2.2
greenlet==3.1.1
h11==0.14.0
httpcore==1.0.7
httpx==0.28.0
idna==3.10
itsdangerous==2.2.0
Jinja2==3.1.4
jiter==0.8.0
jsonify==0.5
logging==0.4.9.6
Mako==1.3.7
MarkupSafe==3.0.2
openai==1.55.3
pydantic==2.10.2
pydantic_core==2.27.1
python-dotenv==1.0.1
python-owasp-zap-v2.4==0.1.0
pytz==2024.2
requests==2.32.3
setuptools==75.6.0
six==1.16.0
sniffio==1.3.1
SQLAlchemy==2.0.36
tqdm==4.67.1
typing_extensions==4.12.2
tzlocal==5.2
urllib3==2.2.3
validators==0.34.0
visitor==0.1.3
Werkzeug==3.1.3
WTForms==3.2.1
zaproxy==0.3.2
zope.interface==7.2
